[NG] Datagrid: switch to flexbox layout
We moved from `table-layout: auto`, which was very useful for smartly sizing columns but far too strict to allow all the features we are planning, to a flexbox layout.

Because of this, we have to manually handle column sizes, which bleeds quite a lot into our Typescript code. To make sure this is done as cleanly as possible, I wrote all the sizing logic in separate providers and classes.

One important pattern to note is the use of the renderer to directly style elements and add classes to them, rather than use more "natural" style and host binding with Angular. The reason is that said bindings introduce watchers and are factored in change detection, which can severely impact the performance of our datagrid. This is why I went with custom event-based size recomputation, to make sure we always did the strict minimum of work, and Angular could simply ignore it during change detection.

This refactor allows 2 nice features out of the box:
1) Datagrids can now scroll easily if you restrict them to their container. The header and footer stay fixed, only the body containing the rows scrolls. It's as simple as setting the height of the datagrid to a specific value, or making it 100% to fill its container.
2) Specifying a header width through CSS or `[style.width]` binding locks the corresponding column size and makes it non-flexible. No need to duplicate that information on the cells, just the header is enough.

Signed-off-by: Eudes Petonnet-Vincent <epetonnetvince@vmware.com>