[NG] Datagrid: cross-browser compatibility
This commit handles browser inconsistencies with flexbox.
It also addresses one obvious but very nasty behavior on OS that actually take space in an element to display a scrollbar (pretty much anything non Apple). In that case, the datagrid body would have less space than the header due to the introduction of the scrollbar, resulting in a misalignment of columns.

Signed-off-by: Eudes Petonnet-Vincent <epetonnetvince@vmware.com>